+++ packages/plugin-spoiler/web/src/tiptap/tiptap.tsx
@@ -1,14 +1,16 @@
 import React from 'react';
+import { CreateRicosExtensions } from 'wix-rich-content-common';
 import { BlockSpoilerComponent } from '..';
 
-export const tiptapExtensions = [
+export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
   {
     type: 'extension' as const,
-    createConfig: () => ({
+    createExtensionConfig: () => ({
       name: 'spoiler',
       priority: 10,
+      defaultOptions,
       addNodeViewHOC: () => ({
-        nodeTypes: [],
+        nodeTypes: ['image'],
         nodeViewHOC: Component => {
           // should use the new api containerData
           return props => {