+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -40,6 +40,8 @@ export interface RicosProps {
   textAlignment?: TextAlignment;
   onAtomicBlockFocus?: onAtomicBlockFocus;
   experiments?: AvailableExperiments;
+  iframeSandboxDomain?: string;
+  textWrap?: boolean;
   /* Changes to this interface should also be reflected in the API docs */
 }