+++ packages/ricos-content/web/lib/translate.ts
@@ -1,0 +1,2 @@
+export { toTranslatables } from '../src/TranslatableAPI/extract/toTranslatables';
+export { translateContent } from '../src/TranslatableAPI/apply/translateContent';