+++ packages/tiptap-editor/web/src/extensions/extension-link/index.ts
@@ -1,0 +1,5 @@
+import { linkMark, linkExtension } from './link';
+
+export const createLink = () => {
+  return [linkMark, linkExtension];
+};