+++ packages/tiptap-editor/web/src/extensions/extension-styles/index.tsx
@@ -1,40 +1,42 @@
 import { getComponentStyles } from './styles';
-import { createRicosGenericExtensionConfig } from '../../extensions-creators/extension';
 import React from 'react';
 
 const name = 'styles';
 
-export const createStylesConfig = () =>
-  createRicosGenericExtensionConfig({
-    type: 'extension',
-    createConfig: () => {
-      return {
-        name,
-        priority: 10,
+const StylesHOC = Component => {
+  const Styles = props => {
+    const { context, componentData, isFocused } = props;
+    const { isMobile, theme } = context;
+    const componentStyles = getComponentStyles({
+      componentData,
+      theme,
+      isFocused,
+      isMobile,
+    });
 
-        addNodeViewHOC() {
-          return {
-            nodeTypes: ['*'],
-            nodeViewHOC: Component => {
-              return props => {
-                const { context, componentData, isFocused } = props;
-                const { isMobile, theme } = context;
-                const componentStyles = getComponentStyles({
-                  componentData,
-                  theme,
-                  isFocused,
-                  isMobile,
-                });
+    return (
+      <div className={Object.values(componentStyles).join(' ')}>
+        <Component {...props} />
+      </div>
+    );
+  };
 
-                return (
-                  <div className={Object.values(componentStyles).join(' ')}>
-                    <Component {...props} />
-                  </div>
-                );
-              };
-            },
-          };
-        },
-      };
-    },
-  });
+  return Styles;
+};
+
+export const createStylesConfig = () => ({
+  type: 'extension',
+  createExtensionConfig: () => {
+    return {
+      name,
+      priority: 30,
+      defaultOptions: {},
+      addNodeViewHOC() {
+        return {
+          nodeTypes: ['*'],
+          nodeViewHOC: StylesHOC,
+        };
+      },
+    };
+  },
+});